requirements.txt
Django==1.10
Pillow==5.0.0
requests==2.18.4
djangorestframework==3.7.7
lxml==4.1.1
psycopg2==2.7.4
coverage==4.5.1